+++ src/PLlib/types.hpp
@@ -10,5 +10,5 @@ typedef short int16_t;
 typedef int int32_t;
 typedef long long int64_t;
 
-typedef unsigned int size_t;
+typedef unsigned long long size_t;
 typedef unsigned long long uintptr_t;
